lscd_calculation_for_mutant_models/requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
editor==1.6.6
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.4
flann==1.6.13
fonttools==4.51.0
fsspec==2024.3.1
idna==3.7
imageio==2.34.1
importlib-metadata==7.1.0
importlib-resources==6.4.0
inquirer==3.2.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==8.6.1
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
kiwisolver==1.4.5
lazy-loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pyflann==1.6.14
pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
pyzmq==26.0.2
readchar==4.0.6
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
runs==1.2.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.4.0
tifffile==2023.7.10
tinycss2==1.2.1
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
triton==2.2.0
typing-extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
xmod==1.8.1
yarg==0.1.9
zipp==3.18.1
